+++ src/main/java/kr/co/teacherforboss/service/boardService/BoardQueryService.java
@@ -1,9 +1,7 @@
 package kr.co.teacherforboss.service.boardService;
 
-import kr.co.teacherforboss.web.dto.BoardRequestDTO;
 import kr.co.teacherforboss.web.dto.BoardResponseDTO;
 
 public interface BoardQueryService {
     BoardResponseDTO.GetPostDTO getPost(Long postId);
-    BoardResponseDTO.GetPresignedUrlDTO getPresignedUrl(BoardRequestDTO.GetPresignedUrlDTO request);
 }
