+++ src/main/java/kr/co/teacherforboss/web/dto/BoardRequestDTO.java
@@ -31,19 +31,4 @@ public class BoardRequestDTO {
         List<String> hashtagList;
     }
 
-    @Getter
-    @AllArgsConstructor
-    @NoArgsConstructor
-    @Builder
-    public static class GetPresignedUrlDTO {
-
-        @NotNull(message = "이미지 저장 목적은 필수 입력값입니다.")
-        String type;
-
-        @NotNull(message = "고유번호는 필수 입력값입니다.")
-        Long id;
-
-        @NotNull(message = "저장할 이미지 개수는 필수 입력값입니다.")
-        Integer imageCount;
-    }
 }
